Relacion en BBDD users y divisions completa

+++ database/migrations/2021_07_18_231841_update_users_table_with_divisions_references.php
@@ -14,7 +14,7 @@ class UpdateUsersTableWithDivisionsReferences extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('division_id')->nullable();
+            $table->unsignedBigInteger('division_id')->nullable();
             $table->foreign('division_id')
                     ->references('id')->on('divisions')
                     ->onDelete('set null');
